Little clean up after classes
