+++ HoneyBadgers.Logic/Data.cs
@@ -36,6 +36,7 @@ namespace HoneyBadgers.Logic
             catch (Exception e)
             {
                 Console.WriteLine($"An error has occurred while reading Users records. Ex: {e.Message}");
+                //TODO: jak pominąć console.write by dostać uniwersalną informacje o błędzie?
                 throw;
             }
         }
